Implemented the execution of a custom __check__ function, when creating, updating or saving a model.
Details:
User may define a __check__(self,data) method for Node-s and Relationship-s.
Here the user may check the coherence of the supplied data. "data" is a dictionary of the defined properties.
This function has no return value but may raise an exception.

The method is run when model is saved to the database (_create,_update,save).
It is up to the user to be prepared for the exceptions.